Splits fitting and plotting modules into fitting.py and plotting.py. Adds helper_functions.py. Adds prepare_cos_fit_dict to fitting.py.
